requirements.txt
Chameleon==2.15
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.19
PasteDeploy==1.5.2
Pygments==1.6
SQLAlchemy==0.9.4
WTForms==1.0.5
WebOb==1.3.1
argparse==1.2.1
oauthlib==0.6.1
pbkdf2==1.3
psycopg2==2.5.2
pyramid==1.4.5
pyramid-debugtoolbar==2.0.2
pyramid-jinja2==1.10
pyramid-mako==0.3.1
pyramid-tm==0.7
python-openid==2.2.5
python-social-auth==0.1.23
repoze.lru==0.6
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
transaction==1.4.3
translationstring==1.1
venusian==1.0a8
waitress==0.8.8
wsgiref==0.1.2
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.4
